+++ src/components/landing_page_v2/flexible_feature_selection.tsx
@@ -6,8 +6,8 @@ import {
   LinearGradientText,
   LinearTextSize,
   TextAlign,
-} from '@/components/landing_page_neo/linear_gradient_text';
-import { LandingButton } from '@/components/landing_page_neo/landing_button';
+} from '@/components/landing_page_v2/linear_gradient_text';
+import { LandingButton } from '@/components/landing_page_v2/landing_button';
 import { LuBellPlus } from 'react-icons/lu';
 
 const FlexibleFeatureIcon: React.FC<{ feature: string; size?: number; halo?: boolean }> = ({
