+++ src/components/landing_page_v2/frosted_glass_block.tsx
