+++ src/components/landing_page_v2/landing_footer.tsx
