+++ src/components/landing_page_v2/landing_navbar.tsx
@@ -3,8 +3,8 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { BsFillRocketTakeoffFill } from 'react-icons/bs';
 import useOuterClick from '@/lib/hooks/use_outer_click';
-import LandingI18n from '@/components/landing_page_neo/landing_i18n';
-import { LandingButton } from '@/components/landing_page_neo/landing_button';
+import LandingI18n from '@/components/landing_page_v2/landing_i18n';
+import { LandingButton } from '@/components/landing_page_v2/landing_button';
 import version from '@/lib/version';
 import { useTranslation } from 'next-i18next';
 import { ISUNFA_ROUTE } from '@/constants/url';
