+++ src/components/landing_page_v2/linear_gradient_text.tsx
